feat: modularize feature of showing total working time
